Add CommandSpec
The CommandSpec allows the user to specify the file the command should execute and the args to provide when executing the file. The name will be used to set the file name in the outdir.

Since this is templatable this commit also updates the template and validate traits for the structs in the command.rs file.