Update CommandSection to execute CommandSpec
CommandSection is updated to execute a command specified using a CommandSpec. This creates the command file in the output directory, makes sure it is executable and then executes.

A set permissions method in the context is created to help.

A CLI test is created to check that commands can be executed from a spec.