RR-111 Support writing multiple files from providers
In order to support writing multiple files we are adding a second trait
for file providers that will be auto-implemented for every type that
implements AsUtf8FileContent. This allows us to keep writing the
relatively simple implementations for providers that we already have if
all we need is to write a single file, while also supporting directly
implementing the new ResolveAndWrite trait to handle things like writing
multiple files or modifying permissions.

In order to be able to add providers that _only_ implement the
ResolveAndWrite we need to update the FileProvider enum to implement it
directly instead of AsUtf8FileContent. Doing so is going to break the
existing tests that expect to be able to call the try_get_file_content
method so that is going to need to be handled first.