RR-113 Pull output from commands via a file
In working on RR-101 and RR-106 (writing the router-scale test plans) we
found that pulling the setup output from a command via stdout ended up
being overly restrictive. It forces the user to keep stdout completely
clean other than the required JSON object which caught us out multiple
times while we tried to get the router-scale scripts working under RTF.

Instead we are taking the approach used by GitHub Actions for
"returning" output from a job:
  https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/passing-information-between-jobs

There will now always be an environment variable named RTF_OUTPUT that
will contain the absolute path to a file that the command can write its
output to. CommandProvider will read the contents of this file and
then delete it before returning the content as the output of the command.

This commit also updates the existing E2E tests to use this new approach
but most of these are not run in CI as they require API keys. To verify
that they are passing locally for you please refer to the instructions
in the test files found in crates/rtf-cli/tests.