RR-113 Add unit test for output file behaviour
The existing test suite covered the user facing behaviour of the new
output file. This unit test checks that we are correctly reading from
the output file and that we remove it once the command has run. This
should work regardless of whether or not the command wrote to the
outfile (though higher level validation around environment setup will
result in an error if the output there is not what was expected).