Add missing static checks for merge_yaml provider
The merge_yaml file provider wasn't performing any static checks for the
two providers it is wrapping. This meant it was possible to provide
invalid data in one or both of the inner providers (such as an invalid
relative path) and for this to only be caught at execution time rather
than in the template and check phase.