Require Context to provide an HTTP client
The http_client() method on Context was originally written to match the
similar methods for obtaining Apollo and GitHub clients which can return
None if the user does not have API keys available for constructing them.
In the case of the HTTP client however, we can always construct the
client and the only place where we returned None was in the default
implementation used in tests when we were not making use of the client.

This was forcing us to include a user facing error condition of no
client being available when in practice hitting this requires us to be
running with a mock implementation of the Context trait. Updating the
method to return a reference to the client rather than an option
simplifies the API and allows us to remove the error condition.